+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ This project adheres to `Semantic Versioning <http://semver.org/>`__.
 [Unreleased]
 ------------
 
+- Added systemd service
+
 [1.0] - 2016-08-15
 ------------------